+++ util/labels/helpers.go
@@ -18,12 +18,7 @@ limitations under the License.
 package labels
 
 import (
-	"encoding/base64"
-	"fmt"
-	"hash/fnv"
-
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/validation"
 
 	clusterv1 "sigs.k8s.io/cluster-api/api/v1beta1"
 )
@@ -41,30 +36,4 @@ func HasWatchLabel(o metav1.Object, labelValue string) bool {
 		return false
 	}
 	return val == labelValue
-}
-
-// MustFormatValue returns the passed inputLabelValue if it meets the standards for a Kubernetes label value.
-// If the name is not a valid label value this function returns a hash which meets the requirements.
-func MustFormatValue(str string) string {
-	// a valid Kubernetes label value must:
-	// - be less than 64 characters long.
-	// - be an empty string OR consist of alphanumeric characters, '-', '_' or '.'.
-	// - start and end with an alphanumeric character
-	if len(validation.IsValidLabelValue(str)) == 0 {
-		return str
-	}
-	hasher := fnv.New32a()
-	_, err := hasher.Write([]byte(str))
-	if err != nil {
-		// At time of writing the implementation of fnv's Write function can never return an error.
-		// If this changes in a future go version this function will panic.
-		panic(err)
-	}
-	return fmt.Sprintf("hash_%s_z", base64.URLEncoding.WithPadding(base64.NoPadding).EncodeToString(hasher.Sum(nil)))
-}
-
-// MustEqualValue returns true if the actualLabelValue equals either the inputLabelValue or the hashed
-// value of the inputLabelValue.
-func MustEqualValue(str, labelValue string) bool {
-	return labelValue == MustFormatValue(str)
 }
